* moved the LightCTL model to a dedicated file * mplemented common methods for set/get actions * reworked LightLigthness to use common get/set methods
