+++ bluetooth_mesh/models/light/__init__.py
@@ -1,3 +1,4 @@
 # pylint: disable=W0614
 
-from .light_lightness import *
+from .lightness import *
+from .ctl import *
